requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
PyJWT==2.8.0
passlib==1.7.4
python-dotenv==1.0.1

#TEST
pytest==8.2.2

# TOOLS
alembic==1.13.1
